Move backup to main and only backup when idle for a second
