+++ main/modules/files.js
@@ -1,8 +1,8 @@
 const path = require('path')
 const fs = require('fs')
 const Store = require('electron-store')
-const { app } = require('electron')
 const log = require('electron-log')
+const { app } = require('electron')
 
 const { t } = require('plottr_locales')
 
@@ -12,6 +12,7 @@ const { emptyFile, tree, SYSTEM_REDUCER_KEYS } = require('pltr/v2')
 const { openProjectWindow } = require('./windows/projects')
 const { shell } = require('electron')
 const { broadcastToAllWindows } = require('./broadcast')
+const { saveBackup } = require('./backup')
 
 const TMP_PATH = 'tmp'
 const TEMP_FILES_PATH = path.join(app.getPath('userData'), 'tmp')
@@ -30,6 +31,29 @@ function saveFile(filePath, jsonData) {
   } else {
     fs.writeFileSync(filePath, JSON.stringify(withoutSystemKeys))
   }
+}
+
+let itWorkedLastTime = true
+
+function autoSave(event, filePath, file) {
+  try {
+    saveFile(filePath, file)
+    // didn't work last time, but it did this time
+    if (!itWorkedLastTime) {
+      itWorkedLastTime = true
+      event.sender.send('auto-save-worked-this-time')
+    }
+  } catch (saveError) {
+    itWorkedLastTime = false
+    event.sender.send('auto-save-error', filePath, saveError)
+  }
+  // either way, save a backup
+  // TODO: move to main modules...
+  saveBackup(filePath, file, (backupError) => {
+    if (backupError) {
+      event.sender.send('auto-save-backup-error', filePath, backupError)
+    }
+  })
 }
 
 function removeFromTempFiles(filePath, doDelete = true) {
@@ -129,6 +153,7 @@ module.exports = {
   TEMP_FILES_PATH,
   tempFilesStore,
   saveFile,
+  autoSave,
   editKnownFilePath,
   removeFromTempFiles,
   removeFromKnownFiles,
